# KDesktopVirt/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KVirtualStage API
  description: |
    Playwright-equivalent desktop automation platform for AI agents.
    
    KVirtualStage provides comprehensive desktop automation capabilities through multiple interfaces:
    - REST API for remote control
    - WebSocket streaming for live desktop viewing
    - Language bindings for Python, Node.js, and C/C++
    - MCP server for AI agent integration
    - CLI and TUI for interactive management
    
    ## Features
    - Natural human-like cursor movement and typing
    - Cross-platform desktop environment support
    - Session recording and playback
    - Workflow automation with multiple steps
    - Real-time performance monitoring
    - Enterprise-grade security and authentication
    
    ## Getting Started
    1. Create a new session with your desired desktop environment
    2. Use automation commands to control the desktop
    3. Optionally start recording to capture interactions
    4. Execute complex workflows with multiple automation steps
    
    ## Authentication
    All API endpoints require valid session authentication. Sessions are created through the `/api/v1/sessions` endpoint.
    
  version: 0.1.0
  contact:
    name: KVirtualStage Team
    url: https://github.com/KooshaPari/KVirtualStage
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.kvirtualstage.dev/v1
...
paths:
  /sessions:
    post:
      summary: Create a new desktop automation session
      description: |
        Creates a new virtual desktop session with the specified configuration.
        The session provides an isolated desktop environment for automation.
      tags:
        - Session Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            examples:
              ubuntu_session:
                summary: Ubuntu desktop session
                value:
                  user_id: "demo_user"
                  session_name: "my_automation_session"
                  desktop_type: "ubuntu"
              kde_session:
                summary: KDE desktop session
                value:
                  user_id: "power_user"
                  session_name: "kde_workflow"
                  desktop_type: "ubuntu-kde"
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  value:
                    success: true
                    data:
                      session_id: "session_20241212_143022"
                      status: "active"
                      vnc_url: "ws://localhost:8080/api/v1/sessions/session_20241212_143022/stream"
                      ws_url: "ws://localhost:8080/api/v1/sessions/session_20241212_143022/stream"
                    error: null
                    timestamp: "2024-12-12T14:30:22Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    get:
      summary: List all active sessions
      description: Retrieves a list of all currently active desktop automation sessions.
      tags:
        - Session Management
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                sessions_list:
                  value:
                    success: true
                    data:
                      - session_id: "session_20241212_143022"
                        user_id: "demo_user"
                        desktop_type: "ubuntu"
                        status: "active"
                        recording_active: false
                      - session_id: "session_20241212_144501"
                        user_id: "power_user"
                        desktop_type: "ubuntu-kde"
                        status: "active"
                        recording_active: true
                    error: null
                    timestamp: "2024-12-12T14:45:30Z"
  /sessions/{session_id}:
    get:
      summary: Get session information
      description: Retrieves detailed information about a specific session.
      tags:
        - Session Management
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Remove/close a session
      description: Terminates and removes a desktop automation session.
      tags:
        - Session Management
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /sessions/{session_id}/recording/start:
    post:
      summary: Start session recording
      description: |
        Begins recording the desktop session to a video file.
        Supports multiple quality levels and formats.
      tags:
        - Recording
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRecordingRequest'
            examples:
              high_quality:
                summary: High quality recording
                value:
                  output_filename: "demo_session.mp4"
                  quality: "high"
              streaming_quality:
                summary: Streaming optimized
                value:
                  output_filename: "live_demo.mp4"
                  quality: "streaming"
      responses:
        '200':
          description: Recording started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      description: Unique session identifier
      schema:
        type: string
        example: "session_20241212_143022"
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the request was successful
        data:
          type: object
          description: Response data (varies by endpoint)
          nullable: true
        error:
          type: string
          description: Error message if request failed
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of the response
      required:
        - success
        - timestamp
    CreateSessionRequest:
      type: object
      properties:
        user_id:
          type: string
          description: User identifier for the session
          example: "demo_user"
        session_name:
          type: string
          description: Optional name for the session
          example: "my_automation_session"
        desktop_type:
          type: string
          enum: ["ubuntu", "ubuntu-xfce", "ubuntu-kde", "centos", "fedora", "arch", "debian"]
          description: Desktop environment type
          default: "ubuntu"
          example: "ubuntu"
      required:
        - user_id
...
    StartRecordingRequest:
      type: object
      properties:
        output_filename:
          type: string
          description: Output video filename
          example: "demo_session.mp4"
        quality:
          type: string
          enum: ["low", "medium", "high", "streaming"]
          description: Recording quality level
          default: "medium"
          example: "high"
      required:
        - output_filename
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            invalid_params:
              value:
                success: false
                data: null
                error: "Invalid parameters: user_id is required"
                timestamp: "2024-12-12T14:30:22Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            session_not_found:
              value:
                success: false
                data: null
                error: "Session not found: session_invalid_id"
                timestamp: "2024-12-12T14:30:22Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            server_error:
              value:
                success: false
                data: null
                error: "Internal server error: failed to initialize desktop environment"
                timestamp: "2024-12-12T14:30:22Z"
...